==> src/burningglass/18_skill_variables.Rmd <==
---
title: "18_skill_variables"
author: "Joanna Schroeder"
date: "9/28/2020"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r message=FALSE, warning=FALSE, include=FALSE, paged.print=FALSE}
# load packages 
for (pkg in c("tidyverse", "data.table", "stringr", "stringi", "mosaic", "dplyr", "readr","gt", 'lubridate')) {
  library(pkg, character.only = TRUE)
}

get_db_conn <-
  function(db_name = "sdad",
           db_host = "postgis1",
           db_port = "5432",
           db_user = Sys.getenv("db_usr"),
           db_pass = Sys.getenv("db_pwd")) {
    RPostgreSQL::dbConnect(
      drv = RPostgreSQL::PostgreSQL(),
      dbname = db_name,
      host = db_host,
      port = db_port,
      user = db_user,
      password = db_pass
    )
  }

con <- get_db_conn()

skill <- DBI::dbGetQuery(con, "SELECT *
FROM bgt_res.skill A
JOIN bgt_res.id_msa_dc B
ON A.id = B.id"
)

DBI::dbDisconnect(con)
```


```{r message=FALSE, warning=FALSE, paged.print=FALSE}
# function for data profiling
data_profiling <- function(df){
summary_table <- tibble(var = names(df),
       variable_type = map_chr(.x = df, .f = function(col) class(x = col)),                 
       num_unique = map_int(.x = df, .f = function(col) length(x = unique(x = col))),
       num_missing = map_int(.x = df, .f = function(col) sum(x = is.na(x = col)))) %>%
  mutate(perc_missing = round(100 * (x = num_missing / nrow(df)), digits = 2L))
return(summary_table)
}
```

## Purpose

We are interested in returning to the BGT tables to add covariates to our analysis. Here we explore the skill variables, which are in the skills table. 

## Profiling

Here we profile the geographic subset of the BGT data for the DC MSA. The skills table is big. For the DC MSA there are there are `r nrow(skill)` rows of data and 650702 unique ids, for an average of `r nrow(skill)/650702` skills per person.

```{r}
#remove duplicate id column
#skill <- skill[,-1]

data_profiling(skill)

head(skill, 20)

skill %>% filter(isbaseline == TRUE) %>% mutate(n = n()) %>% summarise(percent = n/21591463) %>% distinct()

skill %>% mutate(total = n()) %>% filter(isbaseline == TRUE) %>% group_by(skill) %>% mutate(n = n()) %>% summarise(percent = (n/total)*100) %>% distinct() %>% arrange(-percent)

skill %>% mutate(total = n()) %>% filter(issoftware == TRUE) %>% mutate(n = n()) %>% summarise(percent = n/total) %>% distinct()

skill %>% mutate(total = n()) %>% filter(issoftware == TRUE) %>% group_by(skill) %>% mutate(n = n()) %>% summarise(percent = (n/total)*100) %>% distinct() %>% arrange(-percent)

skill %>% mutate(total = n()) %>% filter(isspecialized == TRUE) %>% mutate(n = n()) %>% summarise(percent = n/total) %>% distinct()

skill %>% mutate(total = n()) %>% filter(isspecialized == TRUE) %>% group_by(skill) %>% mutate(n = n()) %>% summarise(percent = (n/total)*100) %>% distinct() %>% arrange(-percent)

skill %>% mutate(total = n()) %>% group_by(clusterfamily) %>% mutate(n = n()) %>% summarise(percent = (n/total) * 100) %>% distinct() %>% arrange(-percent)

skill %>% mutate(total = n()) %>% group_by(cluster) %>% mutate(n = n()) %>% summarise(percent = (n/total) * 100) %>% distinct() %>% arrange(-percent)
```

The skills data is relatively complete because it is mostly imputed data. The cluster and clusterfamily variables are the most incomplete, with about 20% missing. 

About 20% of the skills are designated as "baseline" skills, which appear to be soft skills. About 24% of skills are "software" skills, which includes Microsoft office. About 80% of skills are "specialized". 

The cluster and clusterfamily variables group skills into categories. The most common clusterfamily is information technology, which is not surprising given the nature of the Burning Glass Data. The most commmon cluster is Microsoft Office.

## Comparing nonveterans and veterans

```{r}
bg_vet_demographic <- read.csv("~/git/DSPG2020/career/data/02_bg_vet_demographic.csv")

ids <- bg_vet_demographic$id

skill <- skill %>% mutate(veteran = ifelse(id %in% ids, "veteran", "nonveteran")) 

skill %>% group_by(veteran) %>% mutate(total = n()) %>% group_by(veteran, clusterfamily) %>% mutate(n = n()) %>% mutate(percent = round(n/total * 100, 2)) %>% select(veteran, clusterfamily, percent) %>% distinct() %>%
  ggplot(aes(x = clusterfamily, y = percent, group = veteran, fill = veteran)) +
  geom_col(position = "dodge") +
  theme(axis.text.x = element_text(angle = 60, hjust = 1)) +
  scale_fill_discrete()

skill %>% group_by(id) %>% mutate(n = n()) %>% arrange(id) %>% ungroup() %>% group_by(veteran) %>% summarise(mean = mean(n), median = median(n))

skill %>% group_by(veteran) %>% mutate(total = n()) %>% 
  group_by(veteran, cluster) %>% mutate(n = n()) %>% 
  mutate(percent = round(n/total * 100, 2)) %>% 
  select(veteran, cluster, percent) %>% distinct() %>% group_by(veteran) %>%
  slice_max(percent, n = 10) %>%
  ggplot(aes(x = cluster, y = percent, group = veteran, fill = veteran)) +
  geom_col(position = "dodge") +
  theme(axis.text.x = element_text(angle = 60, hjust = 1)) +
  coord_flip()

```

```{r}
bg_covariates <- fread("~/git/DSPG2020/career/data/bg_covariates_vadcmd.csv") %>% filter(veteran == "veteran")
bg_job <- fread("~/git/DSPG2020/career/data/bg_job_vadcmd.csv")
all_ids <- bg_covariates %>% filter(veteran == "veteran") %>% pull(id)

get_db_conn <-
  function(db_name = "sdad",
           db_host = "postgis1",
           db_port = "5432",
           db_user = Sys.getenv("db_usr"),
           db_pass = Sys.getenv("db_pwd")) {
    RPostgreSQL::dbConnect(
      drv = RPostgreSQL::PostgreSQL(),
      dbname = db_name,
      host = db_host,
      port = db_port,
      user = db_user,
      password = db_pass
    )
  }

con <- get_db_conn()

cert <- DBI::dbGetQuery(con, paste("SELECT * FROM bgt_res.cert WHERE id IN (", paste(ids, collapse = ", "), ")"))

DBI::dbDisconnect(con)


bg_job <- bg_job %>% filter(id %in% bg_covariates$id)
highest_jobs <- bg_job %>% mutate(job_zone = ifelse(onet_job_zone == 55 | onet_job_zone == 6, 0, onet_job_zone)) %>% group_by(id) %>% mutate(highest_zone = max(job_zone)) %>% distinct(id, highest_zone) %>% left_join(bg_covariates, by = "id")

cert_n <- cert %>% group_by(id) %>% mutate(n = n()) %>% distinct(id, n)

ids <- cert_n$id
missing_cert_n_ids <- as.data.frame(all_ids) %>% rename(id = all_ids) %>% filter(!(id %in% ids)) %>% mutate(n = NA)
dataset_cert_n <- cert_n %>% rbind(missing_cert_type_ids)

highest_jobs <- highest_jobs %>% left_join(dataset_cert_n, by = "id")

highest_jobs %>% group_by(n) %>% mutate(total = n()) %>% group_by(highest_zone, n) %>% mutate(percent = round(n()/total * 100, 2)) %>% distinct(highest_zone, n, percent) %>% arrange(highest_zone) %>%
  ggplot(aes(x = n, fill = factor(highest_zone), y = percent)) +
  geom_col(position = "dodge") +
  scale_fill_manual(values = c("#CBBEB5", uva_color_palette[1], uva_color_palette[2],uva_color_palette[4],  uva_color_palette[6], uva_color_palette[9]), labels = c("Education or Military State", "Zone 1", "Zone 2", "Zone 3", "Zone 4", "Zone 5"), name = "Highest Zone Achieved") +
  #theme(axis.text.x = element_text(angle = 60, hjust=1)) +
  ylab("Percent of Non-Veterans and Veterans") +
  xlab("") + #scale_x_discrete(labels = c("veteran" = "Veteran", "not veteran" =
                                           #"Non-Veteran")) +
  theme(legend.position = "non", text = element_text(size = 12))

cert %>% group_by(name) %>% 
  mutate(count = n()) %>% 
  distinct(name, count) %>% filter(count > 40) %>%
  ggplot(aes(x = reorder(name, count), y = as.integer(count)))+
  geom_col() +
  coord_flip()

highest_jobs %>% left_join(cert, by = "id") %>% group_by(name) %>% 
  mutate(avg_zone = mean(highest_zone), count = n()) %>% filter(avg_zone < 3) %>% distinct(name, avg_zone) %>%
  ggplot(aes(x = reorder(name, avg_zone), y = avg_zone)) +
           geom_col() +
           coord_flip()

has_cert <- cert %>% distinct(id, type) %>% filter(type == "License")
ids <- has_cert$id
missing_has_cert_ids <- as.data.frame(all_ids) %>% rename(id = all_ids) %>% filter(!(id %in% ids)) %>% mutate(type = NA)
has_cert <- has_cert %>% rbind(missing_has_cert_ids)

highest_jobs %>% left_join(has_cert, by = "id") %>% group_by(type) %>% 
  mutate(avg_zone = mean(highest_zone)) %>%
  distinct(type, avg_zone) %>%
  ggplot(aes(x = reorder(type, avg_zone), y = avg_zone)) +
           geom_col() +
           coord_flip()

cert
```
